chap11/requirements.txt
attrs==21.4.0
Automat==20.2.0
bcrypt==3.2.0
boto3==1.21.45
botocore==1.24.45
certifi==2021.10.8
cffi==1.15.0
click==8.1.2
constantly==15.1.0
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Script==2.0.6
Flask-Twisted==0.1.2
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
observable==1.0.3
packaging==21.3
pluggy==1.0.0
protobuf==3.20.0
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.1
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
Twisted==22.4.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
zope.interface==5.4.0
